+++ R/mrs_data_display.R
@@ -278,13 +278,16 @@ stackplot <- function(x, ...) {
 #' @param coil the coil element number to plot.
 #' @param labels add labels to each data item.
 #' @param right_marg change the size of the right plot margin.
+#' @param rest_def_par restore default plotting par values after the plot has 
+#' been made.
 #' @param ... other arguments to pass to the matplot method.
 #' @export
 stackplot.mrs_data <- function(x, xlim = NULL, mode = "re", x_units = NULL,
                                fd = TRUE, col = NULL, x_offset = 0,
                                y_offset = 5, dim = "dyn", x_pos = NULL, 
                                y_pos = NULL, z_pos = NULL, dyn = 1, coil = 1, 
-                               labels = NULL, right_marg = NULL, ...) {
+                               labels = NULL, right_marg = NULL,
+                               rest_def_par = TRUE, ...) {
   
   .pardefault <- graphics::par(no.readonly = T)
   
@@ -439,7 +442,7 @@ stackplot.mrs_data <- function(x, xlim = NULL, mode = "re", x_units = NULL,
         #xlim=xlim, xlab="Frequency (ppm)", ylab="Dynamic", 
         #col=gray.colors(64), ...)
   
-  graphics::par(.pardefault)
+  if (restore_def_par) graphics::par(.pardefault)
 }
 
 #' Plot a slice from a 7 dimensional array